fix: to make the database connection
